refactor: move getClaims to context service, add action roles module to context module
