+++ src/auth/auth.utils.ts
@@ -1,6 +1,5 @@
 import { config } from "dotenv";
 import { JwtPayload, verify } from "jsonwebtoken";
-import { AuthTokens } from "./auth.service";
 
 config();
 
@@ -8,21 +7,6 @@ export interface Claims {
   accessTokenClaims: JwtPayload | null;
   refreshTokenClaims: JwtPayload | null;
 }
-
-interface RequestWithCookies extends Request {
-  cookies?: { auth?: AuthTokens };
-}
-
-export const getClaims = (req: RequestWithCookies) => {
-  const { cookies } = req;
-  const accessTokenClaims = cookies?.auth
-    ? decodeToken(cookies.auth.access_token)
-    : null;
-  const refreshTokenClaims = cookies?.auth
-    ? decodeToken(cookies.auth.refresh_token)
-    : null;
-  return { accessTokenClaims, refreshTokenClaims };
-};
 
 export const decodeToken = (token: string) => {
   try {
